refactor(api-server): move the snapshot writer into harness-config; add live-events index
- the harness-config snapshot writer is module infrastructure, not
  bootstrap wiring
- live-events gains a module index so bootstrap imports the composition
  root like every other module
