+++ packages/api-server/src/modules/harness-config/compose.ts
@@ -1,12 +1,8 @@
 import type { Db } from "db";
-import type {
-  HarnessConfigService,
-  HarnessConfigSnapshotPatch,
-} from "api-server-api";
+import type { HarnessConfigService } from "api-server-api";
 import { createHarnessConfigService } from "./services/harness-config-service.js";
 import { createHarnessConfigSnapshotRepo } from "./infrastructure/snapshot-repo.js";
 import type { RuntimeMutator } from "../runtime-delivery/index.js";
-import { emit, EventType } from "../../events.js";
 
 export function composeHarnessConfigModule(deps: {
   db: Db;
@@ -25,27 +21,5 @@ export function composeHarnessConfigModule(deps: {
       getCapabilities: deps.getCapabilities,
       isSettled: deps.isSettled,
     }),
-  };
-}
-
-export function createHarnessConfigSnapshotWriter(deps: {
-  db: Db;
-  resolveOwner: (agentId: string) => Promise<string | null>;
-}): {
-  merge(
-    agentId: string,
-    patch: HarnessConfigSnapshotPatch,
-    opts: { confirmed: boolean },
-  ): Promise<void>;
-} {
-  const repo = createHarnessConfigSnapshotRepo(deps.db);
-  return {
-    async merge(agentId, patch, opts) {
-      await repo.merge(agentId, patch, opts);
-      if (!opts.confirmed) return;
-      const ownerSub = await deps.resolveOwner(agentId);
-      if (!ownerSub) return;
-      emit({ type: EventType.HarnessConfigChanged, agentId, ownerSub });
-    },
   };
 }
