+++ packages/api-server/src/modules/harness-config/index.ts
@@ -1,4 +1,4 @@
 export { composeHarnessConfigModule } from "./compose.js";
 export { harnessConfigSupported } from "./services/harness-config-service.js";
 export { createHarnessConfigSnapshotRepo } from "./infrastructure/snapshot-repo.js";
-export { createHarnessConfigSnapshotWriter } from "./compose.js";
+export { createHarnessConfigSnapshotWriter } from "./infrastructure/snapshot-writer.js";
