+++ packages/api-server/src/modules/live-events/index.ts
@@ -1,0 +1,1 @@
+export { composeLiveEventsModule, type LiveEventsModule } from "./compose.js";
